Fixed parent app bug.
bug:
As a parent, if all of my children in daycare left a specific classroom (Unit), that Unit and associated Supervisors would still remain in the app’s local DB.  Causing an error pop-up stating: “You do not have permission for this Unit”, because you no longer have a Subject linked to that old Unit.  Hence the API error every time the app tried to sync  supervisor details for a Unit it no longer had access to.

resolution:
Since each sync operation in the parent app updates the Units and Supervisors for those units anyway, it’s safe to delete them locally during a sync operation to download updated content.