Enhanced support in resolving sync block with daily logs due to http error code 409 (conflict).
Due to a small window of time between network drops and app backgrounding, the ACK from the server might not reach the app when sending a daily log.  As the server already contains the published daily log for subject, the HTTP error 409 will be thrown causing the app to be stuck in a constant sync block.  (until now)